requirements.txt
disco-py
requests
